refactor: move files related to Gauge
