+++ giraffe/src/components/Gauge/index.ts
@@ -1,0 +1,1 @@
+export {GaugeLayer} from './GaugeLayer'
